Keep non-incremental achievements with progress <100 locked
For incremental achievements, the expected behavior for progress values
less than 100.0 is that the achievement remains locked (and displays
partial progress).  Non-incremental achievements should be treated the
same way.  By doing so, it can be left to the administrator of the
Google Play app to decide what type is most appropriate for an
achievement, and the developer need only report partial progress as
normal and the correct behavior will still happen.