Issue a warning if an outdated range is used in ReportProgress
Users who are using a previous version of the plugin may be assuming
that ``progress`` values are expected to be in the range of [0,1].  As
this is no longer supported (in favor of [0,100]), any ``progress``
value in [0,1] will now trigger a warning.  Such a value is still
applied, as it may be that the user truly does wish to provide a value
in that range (as would be the case for achievements with more than 100
incremental steps).