Do not allocated a pinned GCHandle
Per MSDN
(https://msdn.microsoft.com/en-us/library/vstudio/a95009h1(v=vs.100).asp
x), an attempt to pin a GCHandle for a non-blittable type should throw
an argument exception. The type System.Delegate is non-blittable.

The version of Mono used by Unity does not throw an exception, but the
IL2CPP scripting backend does throw this exception. Therefore, this
code does not work with IL2CPP.

However, it should not be necessary to pin this GCHandle. A normal mode
handle should ensures that the managed object will not be collected.