Fixing leaderboard loadScores & waiting room.
There was a small logic problem when testing for Valid or ValidButStale.

Also updating SmokeTest to show social or public leaderboard and
remove obsolete call to appstate save enablement.

There was a bug fixed of leaving the RTMP room when the app pauses.  This
broke the waiting room, since the activity pauses when the new intent is shown.

Change-Id: I326837b66bf562cd8a961d7a27b96b091bc5cf1b