GPGS Updates for 0.9.30
1. Rebasing Unity Plugin to use the Native SDK CRM API.
2. Added an option to PlayGamesClientConfiguration to require Google+.
3. Added the GetServerAuthCode feature. (NOT PUBLICLY DOCUMENTED.)

Previously, the Unity Plugin was using a wrapper around GmsCore to call the CRM
API. This meant that the CRM API was not supported on iOS. With new versions of
the Native Games SDK, the CRM API is now available to all platforms.

In addition, new Android-only auth functions from the Native SDK (cl/109742544,
cl/109741497) are now exposed in the Unity plugin.

Bug: 25476682, 25978638
Change-Id: Id2ccb9dd88b3ccee8684d5fb37aea07289993957