Updating with Play Games permission changes.
The GetAccess/IdToken APIs are deprecated and should be replaced with
PlayGamesPlayform.GetServerAuthCode().

The email address is still using Google Plus, so the configuration dialog
has changed to allow the developer to require Google Plus access.  This adds
additional permission requirements that the player must consent to, so it should
be used only if absolutely needed.

Change-Id: I54f4a754c42e1b6664872f21e470b83901f04307