Add currentSteps, totalSteps, and isIncremental to PlayGamesAchievement.
PlayGamesPlatform.LoadAchievements returns an array of PlayGamesAchievements.
This object extends Unity's IAchievement, which requires a percentCompleted
field (e.g. 20.0). However, the underlying Google Play achievement actually uses
two values to track progress, currentSteps and totalSteps (e.g. 1 out of 5
represents currentSteps=1 and totalSteps=5). If that fraction is not easily
represented by decimals (for instance, 1 out of 3), percentageCompleted will not
be precise. This CL changes the PlayGamesAchievement class to include the number
of steps in addition to a percentage. It also adds a isIncremental property to
return the type of achievement (incremental or standard).

Bug: 26827811
Change-Id: Ifb2240f40d5e65a823c08613a4236932893da23a