Fixing multiplayer.IsConnectedToRoom() and NO_GPGS compiling.
There is a bug in Play Game Services causing Participant.IsConnectedToRoom() to
always be false.  This was causing problems when setting up a new Realtime game.
The fix is to check the status of the participant and returns true if the
status is JOINED.

Also fixed compiling with NO_GPGS is true, which was not cleaning up the xocode
project correctly.

Change-Id: Ia142716c37be9447269e8856a51752528a1302bb