Fixing using JavaProxy objects on non-main thread.
Unity IL2CPP does not handle JNI proxied objects on the non main thread.
This fix moves the callback into the main thread on the Java side, and
also adds caching of the class object in case there are classloader issues.

Also added missing dependency of play-services-auth which is needed for
the token and email API calls.

Finally, minor tweak to the Minimal sample to log the access token better.

Change-Id: I4f35aa383ab28c2f913dd1d5befa05ae67cef876