Removing iOS support.
This is primarily done to avoid the JarResolver pulling in GPGS via
Cocoapods during an iOS build, even if NO_GPGS is defined.

This forces NO_GPGS to be defined for iOS all the time.  A later change
will clean up the #if (UNITY_ANDROID || (UNITY_IPHONE && !NO_GPGS)) to
be just ANDROID.

Change-Id: If6196c77f5acd340e0061d704eecde3d6b0e88bf